requirements.txt
replicate
numpy
pycocotools
requests
Pillow